+++ src/lib/directives/paginable-table-error.directive.ts
@@ -1,9 +1,0 @@
-import { Directive, TemplateRef } from '@angular/core';
-
-@Directive({
-	selector: '[errorTpt], [paginableTableError]',
-	standalone: true
-})
-export class PaginableTableErrorDirective {
-	constructor(public template: TemplateRef<any>) {}
-}