+++ src/lib/directives/paginable-table-loading.directive.ts
@@ -1,9 +1,0 @@
-import { Directive, TemplateRef } from '@angular/core';
-
-@Directive({
-	selector: '[loadingTpt], [paginableTableLoading]',
-	standalone: true
-})
-export class PaginableTableLoadingDirective {
-	constructor(public template: TemplateRef<any>) {}
-}