+++ src/lib/interfaces/index.ts
@@ -1,6 +1,7 @@
 export * from './column-filter-event';
 export * from './filter-change-event';
 export * from './item-click-event';
+export * from './paginable-state';
 export * from './paginable-action-button';
 export * from './paginable-table-config';
 export * from './paginable-table-dropdown';