+++ src/lib/services/paginable.service.ts
@@ -1,10 +1,10 @@
 import { Inject, Injectable, Optional } from '@angular/core';
 import { DEFAULT_PAGINABLE_CONFIG } from '../constants/defaults';
 import { PaginableTableConfig } from '../interfaces/paginable-table-config';
-import { mergeDeep } from '../utils';
+import { mergeDeep } from 'ng-hub-ui-utils';
 import { PaginableConfigService } from './paginate-config.service';
 
-@Injectable()
+@Injectable({ providedIn: 'root' })
 export class PaginableService {
 	config!: Required<PaginableTableConfig>;